[REF] web,*: calendar: move hooks to their own file
*hr_holidays

This commit regroups all hooks in a new hooks/ folder, and moves
each hook in its own file. It also removes the useClickHandler
hook that was quite small, specific, and used only once, so that
logic is now directly inline in the component needed it.

Followup of task-4510549
