+++ addons/web/static/src/views/calendar/hooks/square_selection_hook.js
